Fix mysql issue reported by Steve.
He was seeing an error message complaining about duplicate
LocalMysqlServer entries.

When I first added the dependency to the Mysql.Web nuget package in
https://github.com/GoogleCloudPlatform/getting-started-dotnet/commit/db9b167d2dde5343e049f062867eeac070d50650
it added a <remove name="LocalMySqlServer /> element that I couldn't prove
was necessary and cluttered the code, so I removed it.  Turns out, in some
cases, like the case Steve discovered, it is necessary.